pyproject.toml
[project]
name = "notifications"
version = "0.1.0"
description = "LLM-Powered House Price Intelligence System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # MCP
    "mcp[cli]>=1.9.3",
    # ML Models
    "catboost>=1.2",
    "scikit-learn>=1.3",
    "umap-learn>=0.5",
    "category-encoders>=2.6",
    "numpy>=1.26",
    "pandas>=2.1",
    # Kafka
    "aiokafka>=0.10",
    # MLflow
    "mlflow>=2.12",
    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    # Database
    "prisma>=0.15",
    # Cache
    "redis>=5.0",
    # LLM
    "openai>=1.30",
    # Utils
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pydantic>=2.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["services*", "kafka*", "mlflow_utils*", "api*", "pipelines*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
